Move profiler into base
That fixes base <-> profiler dependency loop.
It was possible to break profiler -> base dependency, but overall
it embedding profiler code simplifies its "optional" linking.

Anyways, profiler code unit is empty after preprocessing,
if profiling is not enabled.
